Convert header imports to Framework-style.
Using project-style #import style confuses Xcode because it finds the files in multiple places. We always search in the copied header files, because they are all grouped together, unlike the source headers, which are sometimes in sub-folders. Imports in the sources must match those in the headers themselves, or Xcode/clang will pull in different copies and complain about duplicate definitions.